requirements.txt
arrow==0.4.2
certifi==2018.10.15
chardet==3.0.4
ics==0.4
idna==2.7
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.20.0
six==1.11.0
todoist-python==8.1.0
urllib3==1.24
wincertstore==0.2
